pyproject.toml
[project]
name = "alibaba-dataset-analysis"
version = "0.1.0"
description = "Analysis pipeline for Alibaba microservices traces"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "dask[dataframe]>=2024.1.0",
  "distributed>=2024.1.0",
  "pyarrow>=14.0.1",
  "pandas>=2.1.4",
  "numpy>=1.26.2",
  "pyyaml>=6.0.1",
  "networkx>=3.2.1",
  "matplotlib>=3.8.2",
  "plotly>=5.18.0",
  "dask-cuda>=25.12.0",
  "cudf-cu12>=24.12.0",
  "dask-cudf-cu12>=24.12.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
